+++ decidim-admin/app/controllers/decidim/admin/organization_appearance_controller.rb
@@ -2,24 +2,23 @@
 
 module Decidim
   module Admin
-    # Controller that allows managing the user organization.
-    #
-    class OrganizationController < Decidim::Admin::ApplicationController
+    # Controller that allows managing the appearance of the organization.
+    class OrganizationAppearanceController < Decidim::Admin::ApplicationController
       layout "decidim/admin/settings"
 
       def edit
         authorize! :update, current_organization
-        @form = form(OrganizationForm).from_model(current_organization)
+        @form = form(OrganizationAppearanceForm).from_model(current_organization)
       end
 
       def update
         authorize! :update, current_organization
-        @form = form(OrganizationForm).from_params(params)
+        @form = form(OrganizationAppearanceForm).from_params(params)
 
-        UpdateOrganization.call(current_organization, @form) do
+        UpdateOrganizationAppearance.call(current_organization, @form) do
           on(:ok) do
             flash[:notice] = I18n.t("organization.update.success", scope: "decidim.admin")
-            redirect_to edit_organization_path
+            redirect_to edit_organization_appearance_path
           end
 
           on(:invalid) do